home画面を切り分け
