+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_widget_sample/constant/app_string.dart';
-import 'package:flutter_widget_sample/screen/list_builder.dart';
+import 'package:flutter_widget_sample/screen/home.dart';
 
 void main() {
   runApp(const FlutterWidgetSample());
@@ -16,61 +16,7 @@ class FlutterWidgetSample extends StatelessWidget {
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      home: const HomePage(
-        navBarTitle: AppString.appTitle,
-      ),
-    );
-  }
-}
-
-class HomePage extends StatelessWidget {
-  const HomePage({Key? key, required this.navBarTitle}) : super(key: key);
-
-  final String navBarTitle;
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        centerTitle: true,
-        title: Text(
-          navBarTitle,
-        ),
-      ),
-      body: Center(
-        child: _expansionTile(context),
-      ),
-    );
-  }
-
-  Widget _expansionTile(BuildContext context) {
-    return Column(
-      children: [
-        ExpansionTile(
-          title: const Text(AppString.list),
-          children: [
-            ListTile(
-              title: const Text(AppString.listViewBuilder),
-              onTap: () => Navigator.push(
-                context,
-                MaterialPageRoute<void>(
-                  builder: (context) {
-                    return const ListBuilderView();
-                  },
-                ),
-              ),
-            ),
-            const ListTile(
-              title: Text(AppString.listViewGenerate),
-              // onTap: {},
-            ),
-            const ListTile(
-              title: Text(AppString.listViewSeparated),
-              // onTap: {},
-            ),
-          ],
-        ),
-      ],
+      home: const Home(),
     );
   }
 }
